utils: move vec_rotate to math dir.
